# enable pyogrio to read natural earth data