--- a/.github/workflows/ci-additional.yaml
+++ b/.github/workflows/ci-additional.yaml
@@ -26,6 +26,8 @@ jobs:
             python-version: "3.7"
           - env: "min-all-deps"
             python-version: "3.7"
+          - env: "min-all-deps"
+            python-version: "3.8"
           - env: "shapely-2"
             python-version: "3.10"
     steps: