--- a/.github/workflows/ci-additional.yaml
+++ b/.github/workflows/ci-additional.yaml
@@ -105,4 +105,4 @@ jobs:
 
       - name: minimum versions policy
         run: |
-          python ci/min_deps_check.py ci/requirements/py37-min-all-deps.yml
+          python ci/min_deps_check.py ci/requirements/py37-min-all-deps.yml ci/requirements/py38-min-all-deps.yml