--- a/ci/requirements/environment.yml
+++ b/ci/requirements/environment.yml
@@ -10,6 +10,7 @@ dependencies:
   - packaging
   - pooch
   - pygeos
+  - pyogrio
   - rasterio
   - setuptools
   - xarray