--- a/ci/requirements/py37-min-all-deps.yml
+++ b/ci/requirements/py37-min-all-deps.yml
@@ -8,10 +8,11 @@ dependencies:
   - geopandas=0.7
   - matplotlib-base=3.2
   - numpy=1.17
-  - pandas=1.2
   - packaging=21.3
+  - pandas=1.2
   - pooch=1.2
   - pygeos=0.9
+  - pyogrio=0.3
   - rasterio=1.1
   - setuptools=40.4
   - shapely=1.7