--- a/ci/requirements/py37-min-all-deps.yml
+++ b/ci/requirements/py37-min-all-deps.yml
@@ -12,7 +12,6 @@ dependencies:
   - pandas=1.2
   - pooch=1.2
   - pygeos=0.9
-  # - pyogrio=0.3 # requires py3.8+
   - rasterio=1.1
   - setuptools=40.4
   - shapely=1.7