--- a/ci/requirements/py38-min-all-deps.yml
+++ b/ci/requirements/py38-min-all-deps.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - nodefaults
 dependencies:
-  - python=3.7
+  - python=3.8
   - cartopy=0.17
   - geopandas=0.7
   - matplotlib-base=3.2
@@ -12,7 +12,7 @@ dependencies:
   - pandas=1.2
   - pooch=1.2
   - pygeos=0.9
-  # - pyogrio=0.3 # requires py3.8+
+  - pyogrio=0.3 # requires py3.8+
   - rasterio=1.1
   - setuptools=40.4
   - shapely=1.7