--- a/ci/requirements/py38-min-all-deps.yml
+++ b/ci/requirements/py38-min-all-deps.yml
@@ -4,18 +4,18 @@ channels:
   - nodefaults
 dependencies:
   - python=3.8
-  - cartopy=0.17
-  # - geopandas=0.7 # not available in py38
-  - geopandas=0.9
-  - matplotlib-base=3.2
-  - numpy=1.17
+  - cartopy=0.20
+  - geopandas=0.8
+  - matplotlib-base=3.5
+  - numpy=1.21
   - packaging=21.3
   - pandas=1.2
   - pooch=1.2
-  - pygeos=0.9
-  - pyogrio=0.3 # requires py3.8+
-  - rasterio=1.1
-  - shapely=1.7
+  - pygeos=0.12
+  # pyogrio makes environment solving difficult
+  - pyogrio=0.3
+  - rasterio=1.2
+  - shapely=1.8
   - xarray=0.15
 # for testing
   - pytest