<!-- Feel free to remove check-list items aren't relevant to your change -->

 - [x] Closes #404
 - [x] Tests added
 - [x] Passes `isort . && black . && flake8`
 - [ ] Fully documented, including `whats-new.rst`

This should speed up the natural earth data shapefiles - even after #405 (which already sped up the tests from 60s to 15s). Not this one should bring it down from 15s to 2s (assuming the data is already downloaded).